editor-preview: move the LSP child-process connector
