+++ internal/editor-preview/child_process.rs
@@ -1,0 +1,129 @@
+// Copyright © SixtyFPS GmbH <info@slint.dev>
+// SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-Slint-Royalty-free-2.0 OR LicenseRef-Slint-Software-3.0
+
+use std::cell::RefCell;
+
+use i_slint_live_preview::protocol::{LspToPreviewMessage, PreviewTarget, PreviewToLspMessage};
+use tokio::{
+    io::{AsyncBufReadExt, AsyncWriteExt},
+    sync::mpsc,
+    task::JoinHandle,
+};
+
+struct ChildProcessLspToPreviewInner {
+    communication_handle: JoinHandle<Result<(), String>>,
+    to_child_sender: mpsc::UnboundedSender<String>,
+}
+
+pub struct ChildProcessLspToPreview {
+    inner: RefCell<Option<ChildProcessLspToPreviewInner>>,
+    preview_to_lsp_channel: mpsc::UnboundedSender<PreviewToLspMessage>,
+}
+
+impl ChildProcessLspToPreview {
+    pub fn new(preview_to_lsp_channel: mpsc::UnboundedSender<PreviewToLspMessage>) -> Self {
+        Self { inner: RefCell::new(None), preview_to_lsp_channel }
+    }
+
+    fn preview_is_running(&self) -> bool {
+        self.inner.borrow().as_ref().map(|i| !i.communication_handle.is_finished()).unwrap_or(false)
+    }
+
+    fn start_preview(&self) -> crate::Result<()> {
+        if let Some(inner) = self.inner.borrow_mut().take() {
+            inner.communication_handle.abort();
+        }
+
+        let mut child = tokio::process::Command::new(
+            std::env::current_exe().expect("Could not find executable name of the slint-lsp"),
+        )
+        .args(["live-preview", "--remote-controlled"])
+        .stdin(std::process::Stdio::piped())
+        .stdout(std::process::Stdio::piped())
+        .spawn()?;
+
+        tracing::debug!("Preview process spawned (PID {:?})", child.id());
+
+        let from_child = child.stdout.take().expect("Child has no stdout");
+        let mut to_child = child.stdin.take().expect("Child has no stdin");
+
+        let channel = self.preview_to_lsp_channel.clone();
+
+        let preview_to_lsp_channel = self.preview_to_lsp_channel.clone();
+
+        let communication_handle = tokio::spawn(async move {
+            let reader = tokio::io::BufReader::new(from_child);
+            let mut lines = reader.lines();
+            while let Some(line) = lines.next_line().await.map_err(|e| e.to_string())? {
+                if let Ok(message) = serde_json::from_str(&line) {
+                    channel.send(message).map_err(|e| e.to_string())?;
+                }
+            }
+
+            let exit_status = child.wait().await.map_err(|e| e.to_string());
+
+            if exit_status.map(|exit_status| !exit_status.success()).unwrap_or(true) {
+                let message =
+                    "The Slint live preview crashed! Please open a bug on the [Slint bug tracker](https://github.com/slint-ui/slint/issues)."
+                        .to_string();
+                tracing::error!("{message}");
+
+                let _ = preview_to_lsp_channel.send(PreviewToLspMessage::SendShowMessage {
+                    message: lsp_types::ShowMessageParams {
+                        typ: lsp_types::MessageType::ERROR,
+                        message,
+                    },
+                });
+            }
+            Ok(())
+        });
+
+        let (to_child_sender, mut to_child_receiver) = mpsc::unbounded_channel::<String>();
+        tokio::spawn(async move {
+            while let Some(mut msg) = to_child_receiver.recv().await {
+                msg.push('\n');
+                if let Err(err) = to_child.write_all(msg.as_bytes()).await {
+                    tracing::error!("Failed writing to preview child process: {err}");
+                    break;
+                }
+            }
+        });
+
+        *self.inner.borrow_mut() =
+            Some(ChildProcessLspToPreviewInner { communication_handle, to_child_sender });
+
+        Ok(())
+    }
+}
+
+impl Drop for ChildProcessLspToPreview {
+    fn drop(&mut self) {
+        if let Some(inner) = self.inner.borrow_mut().take() {
+            let message = serde_json::to_string(&LspToPreviewMessage::Quit).unwrap();
+            let _ = inner.to_child_sender.send(message);
+        }
+    }
+}
+
+impl crate::LspToPreview for ChildProcessLspToPreview {
+    fn send(&self, message: &LspToPreviewMessage) {
+        if self.preview_is_running() {
+            let mut inner = self.inner.borrow_mut();
+            let inner = inner.as_mut().unwrap();
+            let Ok(message) = serde_json::to_string(message) else {
+                tracing::debug!("Failed to serialize message to preview");
+                return;
+            };
+            let _ = inner.to_child_sender.send(message);
+        } else if let LspToPreviewMessage::ShowPreview(_) = message {
+            tracing::debug!("Starting preview process");
+            self.start_preview().unwrap();
+        } else {
+            tracing::warn!("Preview not running, dropping message: {:?}", message);
+        }
+    }
+
+    fn preview_target(&self) -> PreviewTarget {
+        PreviewTarget::ChildProcess
+    }
+}
